sub-packages/bionemo-testing/requirements.txt
boto3
email-validator
platformdirs
pooch
pydantic>=2.7.0
pytest
pyyaml
tqdm
